pyproject.toml
[tool.poetry]
name = "prompt-evals"
version = "0.1.0"
description = ""
authors = ["margostino <maj.dagostino@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.12.0"
streamlit = "^1.31.1"
watchdog = "^4.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
